fixed issues with motors not moving and added a feature and began another feature
1. added actual dyanic memory constructors for the servos, this allowed
activation of the servos
2. removed a 10 ms delay that was between the display on/off for some reason
3. moved the 1ms time delay in the loop to be between dipslay power on/off,
to ensure proper brightness
4. added remote debounce, now there is a small timer that blocks repeat
inputs from a button press that is too long. This operates through an
irTimer class, look for the irDebouncer object
5. started with adding memory capabilities for the memory holder. There
is now two new variables in the numberholder class that act as a
previous 2 digit memory. This might be better implimented with a very
large array with moving frame.